+++ src/Catalog/CatalogService.php
@@ -11,22 +11,18 @@ use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\RequestOptions;
 use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Exception\NotFound;
+use Wizaplace\SDK\Exception\ProductNotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use function theodorejb\polycast\to_string;
 
-final class CatalogService extends AbstractService
+final class CatalogService extends AbstractService implements CatalogServiceInterface
 {
+    /**
+     * @throws ProductNotFound
+     */
     public function getProductById(string $id) : Product
     {
-        try {
-            $response = $this->client->get("catalog/products/{$id}");
-        } catch (ClientException $exception) {
-            if ($exception->getResponse()->getStatusCode() === 404) {
-                throw new NotFound("Product #{$id} not found.", $exception);
-            }
-
-            throw $exception;
-        }
+        $response = $this->client->get("catalog/products/{$id}");
 
         return new Product($response, $this->client->getBaseUri());
     }
@@ -155,7 +151,7 @@ final class CatalogService extends AbstractService
     /**
      * Report a suspicious product to the marketplace administrator.
      *
-     * @throws NotFound
+     * @throws ProductNotFound
      * @throws SomeParametersAreInvalid
      */
     public function reportProduct(ProductReport $report): void
@@ -173,8 +169,6 @@ final class CatalogService extends AbstractService
             ]);
         } catch (ClientException $e) {
             switch ($e->getCode()) {
-                case 404:
-                    throw new NotFound("Product #{$report->getProductId()} not found", $e);
                 case 400:
                     throw new SomeParametersAreInvalid(to_string($e->getResponse()->getBody()), 400, $e);
             }