+++ CHANGELOG.md
@@ -34,6 +34,20 @@
 
 </details>
 
+## 1.16.0
+
+### New features
+
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\ProductNotFound`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\ReviewsAreDisabled`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\SenderIsAlsoRecipient`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\CompanyHasNoAdministrator`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\CompanyNotFound`
+ - Improve detection of `\Wizaplace\SDK\Favorite\Exception\FavoriteAlreadyExist`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\BasketIsEmpty`
+ - Added `\Wizaplace\SDK\Catalog\CatalogServiceInterface` which is implemented by `\Wizaplace\SDK\Catalog\CatalogService`
+ - Added `\Wizaplace\SDK\Catalog\AbstractCatalogServiceDecorator` which allows to easily decorate any `\Wizaplace\SDK\Catalog\CatalogServiceInterface`
+
 ## 1.15.1
 
  - Fixed type errors on `\Wizaplace\SDK\Catalog\Facet\ListFacetValue::getCount` and `\Wizaplace\SDK\Catalog\Facet\ListFacetValue::getPosition`